% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{A correlation function}
\usage{
correlation(year)
}
\arguments{
\item{year}{Year of interest}
}
\description{
This function returns a correlation matrix for all numeric variables in the data for that year.
}
\examples{
correlation(2005)
}
\keyword{correlation}
