% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playerstats.R
\name{player_stats}
\alias{player_stats}
\title{A Player Assists Function}
\usage{
player_stats(year)
}
\arguments{
\item{year}{Year of interest.}
}
\description{
This function allows you to see which player had the most assists in a specific year
}
\examples{
player_stats(2005)
}
\keyword{assists}
